refactor: update comments, make new folder
